# Fixing performance issues in current BWMF task implementation.