--- a/.script/test
+++ b/.script/test
@@ -6,16 +6,9 @@ if [ ! -d ".tmp" ]; then
     mkdir .tmp
 fi
 
-clean_test_files() {
-  rm	example/bwmf/data/dShard.dat.0 > /dev/null 2>&1
-  rm	example/bwmf/data/dShard.dat.1 > /dev/null 2>&1
-  rm	example/bwmf/data/tShard.dat.0 > /dev/null 2>&1
-  rm	example/bwmf/data/tShard.dat.1 > /dev/null 2>&1
-}
 cleanup() {
   kill $(jobs -p) 2>/dev/null
   rm -r $ETCDTESTDIR
-  clean_test_files
 }
 # cleanup on any kind of exit
 trap "cleanup" SIGINT SIGTERM EXIT