--- a/example/bwmf/kldiv_loss.go
+++ b/example/bwmf/kldiv_loss.go
@@ -1,8 +1,8 @@
 package bwmf
 
 import (
-	"log"
 	"math"
+	"sort"
 
 	pb "github.com/taskgraph/taskgraph/example/bwmf/proto"
 	"github.com/taskgraph/taskgraph/op"
@@ -16,12 +16,23 @@ import (
 //  So actually H is H^T, but it saves code by using identical routine when alternatively optimize over H and W.
 //
 type KLDivLoss struct {
-	V       *pb.SparseMatrixShard
-	W       op.Parameter // blockParameter. An op.Parameter wrapper with map[uint32]*DenseMatrixShard as underlying storage
-	WH      [][]float32  // temporary storage for the intermediate result W*H
-	m, n, k int          // dimensions
+	V       *pb.MatrixShard
+	W       []*pb.MatrixShard
+	m, n, k int // dimensions
 	smooth  float32
-	logger  *log.Logger
+
+	ends []int // for binary search of blockId.
+}
+
+func NewKLDivLoss(v *pb.MatrixShard, w []*pb.MatrixShard, m, n, k int, smooth float32) *KLDivLoss {
+	stt := make([]int, len(w)+1)
+	ost := 0
+	for i, m := range w {
+		stt[i] = ost
+		ost += len(m.Row)
+	}
+	stt[len(w)] = ost
+	return &KLDivLoss{V: v, W: w, m: m, n: n, k: k, smooth: smooth, ends: stt}
 }
 
 // This function evaluates the Kullback-Leibler Divergence given $\mathbf{V} the matrix to fact and $\mathbf{W}$ the fixed factor.
@@ -44,39 +55,33 @@ type KLDivLoss struct {
 //
 func (l *KLDivLoss) Evaluate(param op.Parameter, gradient op.Parameter) float32 {
 	H := param
-	var value float32
+	op.Fill(gradient, 0.0)
+	value := float32(0.0)
+	wi := make([]float32, l.k)
 	for i := 0; i < l.m; i++ {
 		for j := 0; j < l.n; j++ {
-			l.WH[i][j] = 0.0
+			v, _ := l.V.GetRow()[j].At[int32(i)]
+			wh := l.smooth // move away from 0
+
 			for k := 0; k < l.k; k++ {
-				l.WH[i][j] += l.W.Get(i*l.k+k) * H.Get(j*l.k+k)
+				w := l.GetWElem(i, k)
+				wi[k] = w
+				wh += w * H.Get(j*l.k+k)
 			}
 
 			// evaluate element-wise KL-divergence
-			v, ok := l.V.GetRow()[j].At[int32(i)]
-			wh := l.WH[i][j]
-			if ok {
-				value += -v*float32(math.Log(float64(l.smooth+wh))) + wh
-			} else {
-				value += wh
+			value += -v*float32(math.Log(float64(wh))) + wh
+
+			// accumulate to grad vec
+			for k := 0; k < l.k; k++ {
+				gradient.Add(j*l.k+k, wi[k]*(wh-v)/wh)
 			}
 		}
 	}
-	// now, another pass for grad calculation
-	for j := 0; j < l.n; j++ {
-		for k := 0; k < l.k; k++ {
-			grad_index := j*l.k + k
-			gradient.Set(grad_index, 0.0)
-			for i := 0; i < l.m; i++ {
-				grad_val := l.W.Get(i*l.k+k) * (l.WH[i][j] - l.V.GetRow()[j].At[int32(i)]) / l.WH[i][j]
-				gradient.Add(grad_index, grad_val)
-			}
-		}
-	}
-
-	if l.logger != nil {
-		l.logger.Println("Evaluate at param ", param, " returns value ", value, " and gradient ", gradient)
-	}
-
 	return value
+}
+
+func (l *KLDivLoss) GetWElem(row, column int) float32 {
+	b := sort.Search(len(l.ends), func(i int) bool { return l.ends[i] > row }) - 1
+	return l.W[b].Row[row-l.ends[b]].At[int32(column)]
 }