This PR includes modifications and improvements:

* A unified sparse representation of matrix shard. We do not need dense representation any more, instead we use a vec parameter to store param data and copy the values to matrix shard with 0s filtered.
* An improved implementation of KL divergence loss. We now do one pass to evaluate the loss and accumulate the gradient.
* T/D shard are initialized as sparse/random matrices if not loaded.