@BaiGang 2015-04-27
Now the cost of evaluation KL-DIv<V, W*H> is O(m*n*nonzero(W)), where V is mxn, W is mxk and H is kxn.