@5kg 2015-04-27
> Now the cost of evaluation KL-DIv is O(mnnonzero(W)), where V is mxn, W is mxk and H is kxn.

I saw an improvement 122.460s -> 95.141s