@5kg 2015-04-27
I tried to understand the effects of these series of changes.

* with 35054ee59696e27827323e67c84cbb87735f5df0 alone, I saw `457.892s -> 119.28s`
* with the spareness patch 64f4fa3a503fcebf0b536db04b0f95dd481c2764 (no thresholding, no new initialization method): `119.28s -> 150.105s`
* with thresholding: `150.105s -> 134.245s`
* with new initialization method: `134.245s -> 76.420s`, BUT performance get worse (in term of loss `80879.351562 v.s. 48503.832031`)