@5kg 2015-04-27
New profiling result, ![profile](http://i.imgur.com/cIBUfms.png)

map access is now the hot-spot.