@BaiGang 2015-04-30
We got a build error in [the trivis build run](https://travis-ci.org/taskgraph/taskgraph/builds/60640479) for commit 0e111ca:

```
../../colinmarc/hdfs/client.go:4:2: cannot find package "code.google.com/p/goprotobuf/proto" in any of:
	/home/travis/.gimme/versions/go1.4.linux.amd64/src/code.google.com/p/goprotobuf/proto (from $GOROOT)
	/home/travis/gopath/src/code.google.com/p/goprotobuf/proto (from $GOPATH)
The command ".script/test" exited with 1.
```

It seems irrelevant here.