@xiaoyunwu 2015-04-30
@hongchao, can you take a look?

Xiaoyun

On Thu, Apr 30, 2015 at 12:01 PM, Gang Bai <notifications@github.com> wrote:

> We got a build error in the trivis build run
> <https://travis-ci.org/taskgraph/taskgraph/builds/60640479> for commit
> 0e111ca
> <https://github.com/taskgraph/taskgraph/commit/0e111cac223e1e2ba2bbabd7f3cb72c9fb421965>
> :
>
> ../../colinmarc/hdfs/client.go:4:2: cannot find package "code.google.com/p/goprotobuf/proto" in any of:
>     /home/travis/.gimme/versions/go1.4.linux.amd64/src/code.google.com/p/goprotobuf/proto (from $GOROOT)
>     /home/travis/gopath/src/code.google.com/p/goprotobuf/proto (from $GOPATH)
> The command ".script/test" exited with 1.
>
> It seems irrelevant here.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/taskgraph/taskgraph/pull/83#issuecomment-97654166>.
>