@fengjingchao 2015-04-30
No. We don't have that dependency.
Just add any dependency we want to ".script/dep.sh"