@BaiGang 2015-05-05
Next, we will be working on a parallel version of `KLDivLoss.Evaluate()`. @5kg