@fengjingchao 2015-05-05
Is this ready to commit? And can you remove the binary files?