@BaiGang 2015-05-05
I think it is ready to merge. @xiaoyunwu Please take a look if it's in the
right direction. Next we'll be parallelizing the evaluation and
accumulation of loss and gradients in kldiv_loss.go.

As for the binary files, they are actually the marshalled data of matrix
shard for integration test. Shall we generate these files on the fly during
integration test, or just leave it as is?

On Tue, May 5, 2015 at 9:28 AM, Hongchao Deng <notifications@github.com>
wrote:

> Is this ready to commit? And can you remove the binary files?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/taskgraph/taskgraph/pull/83#issuecomment-98899705>.
>