@fengjingchao 2015-05-05
Keep it if you need.
I remove some test generated files in script:
https://github.com/taskgraph/taskgraph/blob/master/.script/test#L10

I think those files should be generated in a dir like "tmp/" to be separated from "data/", and add the dir to gitignore.