@BaiGang 2015-05-05
OK. I'll modify it to generate test data in `tmp/` and push it later.

On Tue, May 5, 2015 at 9:35 AM, Hongchao Deng <notifications@github.com>
wrote:

> Keep it if you need.
> I remove some test generated files in script:
> https://github.com/taskgraph/taskgraph/blob/master/.script/test#L10
>
> I think those files should be generated in a dir like "tmp/" to be
> separated from "data/", and add the dir to gitignore.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/taskgraph/taskgraph/pull/83#issuecomment-98901756>.
>